modified:   u2net/train_utils/train_and_eval.py 	modified:   fcn/__pycache__/my_dataset.cpython-38.pyc 	modified:   fcn/my_dataset.py 	modified:   fcn/train_multi_GPU.py 	modified:   u2net/__pycache__/my_dataset.cpython-38.pyc 	modified:   u2net/my_dataset.py 	modified:   u2net/train_multi_GPU.py 	lraspp/

+++ u2net/train_utils/train_and_eval.py
@@ -6,7 +6,7 @@ import train_utils.distributed_utils as utils
 
 def criterion(inputs, target):
     # weights = torch.linspace(0.1, 1, len(inputs))
-    losses = [F.cross_entropy(inputs[i], target.squeeze().long(), ignore_index=255)
+    losses = [F.cross_entropy(inputs[i], target.squeeze(), ignore_index=255)
               for i in range(len(inputs))]
     total_loss = sum(losses)
